Skapade user ws

+++ src/main/java/WebbAppPVT15Grupp2/Models/User.java
@@ -2,15 +2,15 @@ package WebbAppPVT15Grupp2.Models;
 
 public class User {
 
-    private int id;
+    private long id;
     private String name;
 
-    public User (int id, String name){
+    public User (long id, String name){
             this.id = id;
             this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
